refactor(synapsys): split lint into focused modules
Extract parse-args, pair-scoring, pretool-scoring, broad-triggers,
suggestion, and formatters into lib/lint/ modules. Resolves 8 lint errors
for complexity and max-lines. Public API preserved in synapsys-lint.js.
